requirements.txt
Flask
Flask-SQLAlchemy
marshmallow
marshmallow-sqlalchemy
python-dotenv
psycopg2-binary